Online Disaster Reporting System
I'm developing a web application for the ‘Department of Disaster Management’ that can be used to report disasters (both natural & man made) that occur in the western province. Access to reporter accounts for this application is given to authorized persons from the police stations, “grama niladari office” & Post Office and assigned members from general public.